polygon/tests: move mining bench from tests to polygon/tests
